content: generateTrackInfoPageOtherReleasesList
